refactor: ActionData 패키지 이동 및 toEntity 로 한번에, 유저 사진 정보 임시 추가
