+++ backend/src/main/java/kr/codesquad/todo/dto/response/ActionData.java
@@ -1,30 +1,33 @@
-package kr.codesquad.todo.dto;
+package kr.codesquad.todo.dto.response;
 
 import java.time.LocalDateTime;
 
+import kr.codesquad.todo.domain.Action;
 import kr.codesquad.todo.domain.ActionType;
-import kr.codesquad.todo.dto.response.ActionResponse;
 
 public class ActionData {
 
 	private final String nickname;
+	private final String imageUrl;
 	private final String cardTitle;
 	private final String originCategoryName;
 	private final String targetCategoryName;
 
 	public ActionData(String cardTitle, String originCategoryName, String targetCategoryName) {
+		// 임시 유저 정보 -> 나중에 데이터 받는 부분 옮겨야 할 듯
 		this.nickname = "bruni";
+		this.imageUrl = "https://github-production-user-asset-6210df.s3.amazonaws.com/48724199/254183695-5d025f0a-e616-494d-8ec0-287a279d800f.jpg";
 		this.cardTitle = cardTitle;
 		this.originCategoryName = originCategoryName;
 		this.targetCategoryName = targetCategoryName;
 	}
 
-	public String getNickname() {
-		return nickname;
-	}
-
 	public ActionData(String cardTitle) {
 		this(cardTitle, null, null);
+	}
+
+	public String getNickname() {
+		return nickname;
 	}
 
 	public String getCardTitle() {
@@ -39,14 +42,15 @@ public class ActionData {
 		return targetCategoryName;
 	}
 
-	public ActionResponse toResponse(ActionType actionType) {
-		return new ActionResponse(
-			nickname,
+	public Action toEntity(ActionType actionType) {
+		return new Action(
+			null,
 			actionType.getDescription(),
 			cardTitle,
 			originCategoryName,
 			targetCategoryName,
-			LocalDateTime.now()
+			LocalDateTime.now(),
+				1L
 		);
 	}
 }
